getCanonicalName abstract, makeSignature has receiver
Regarding getCanonicalName(), the previous interface required subclasses to
implement getPackageName() and getSimpleName(), however these methods were
really only used in the construction of getCanonicalName() and not terribly
useful on their own. The problem with this approach is that getCanonicalName()
was therefore required to return a const char*, which could result in either
corrupt memory from referencing a stack variable or leaked memory (depending
on whether strdup was used).

This could be avoided by requiring callers to clean up after themselves by
storing the result of getCanonicalName() in a variable and then freeing it,
however I believe this would result in a lot of leaks since it's easy to
forget this.

For that reason, getCanonicalName(), which is the method most callers actually
care about, now is guaranteed to return a heap pointer, preferably one created
by a preprocessor macro that will construct a package name for you. The method
getSimpleName() still remains, and simply looks for the name portion of the
canonical name. This should be much safer and less leaky.

Likewise, make* methods now require a receiver to store the result.