Add simple test case, JUnit utilities
This is cool -- instead of the silly Success exception, we should now be able
to throw exceptions that JUnit recognizes. They show up nicely in the IDE,
though they are not usually formatted in the same way as JUnit does it
(however this doesn't really make sense for native exceptions anyways).