Move persistence features to subclass
Rather than trying to cram all the persistence stuff into JavaClass
(formerly ClassWrapper), now these features are present in a separate
subclass, NativeObject. The nice thing here is that persist() can be
called implicitly for the user (though it is still public and can be
invoked directly), and there is a corresponding Java class which
provides the destroy() method and nPtr field for the user. MUCH nicer
interface, not sure why I didn't think of this when designing
ClassWrapper.