Make destroy() an abstract method
If destroy is not abstract, then there is a good chance that we will run into
problems when attempting to register the native method for this, because JNI
doesn't let you implicitly register methods only overridden by a subclass.

I'm not terribly fond of this interface, as it requires a bit more boilerplate
code in the subclass than I would prefer, however it's preferable to crashing.